把 Track 相关的模型类整理到一个文件中
