+++ src/model/__init__.py
@@ -2,7 +2,7 @@ from src.model.exercise import Exercise
 from src.model.headers import Headers
 from src.model.route import Route
 from src.model.route_group import RouteGroup
-from src.model.track.track import Track
+from src.model.track import Track
 from src.model.user import User
 
 __all__ = ["Exercise", "Headers", "Route", "Track", "User", "RouteGroup"]
